### 2021-05-04 #3246 @whedon
[PRE REVIEW]: Computer-Aided Generation of N-shift RWS
**Submitting author:** @benjaminbolling (<a href="http://orcid.org/0000-0002-6650-5365">Benjamin Edward Bolling</a>)
**Repository:** <a href="https://github.com/benjaminbolling/RSW" target ="_blank">https://github.com/benjaminbolling/RSW</a>
**Version:** v.1.0.0
