+++ docs/paper.bib
@@ -5,6 +5,7 @@
   title =         {{Efficient generation of rotating workforce schedules}},
   volume =        {118},
   year =          {2002}
+  doi =           {10.1016/s0166-218x(01)00258-x}
 }
 
 @webpage{PythonReference,
