+++ docs/paper.bib
@@ -4,7 +4,7 @@
   pages =         {85--98},
   title =         {{Efficient generation of rotating workforce schedules}},
   volume =        {118},
-  year =          {2002}
+  year =          {2002},
   doi =           {10.1016/s0166-218x(01)00258-x}
 }
 
