### 2021-05-05 a008fb7
Figure 0 cropped and added into paper
