### 2021-05-05 9c02324
Motivation section shortened
