### 2021-05-05 17d8ea8
Motivation section shortened
