### 2021-05-05 952d8de
Motivation section shortened
