### 2021-05-05 5b1ae5a
Motivation section shortened
