### 2021-05-05 cd25439
trying again with the equation figure
