### 2021-05-05 bcbd797
trying again with the equation figure
