### 2021-05-05 0dbcc42
trying again with the equation figure
