### 2021-05-05 bf10d28
trying again with the equation figure
