### 2021-05-05 272898c
trying again with the equation figure
