### 2021-05-05 e773992
trying again with the equation figure
