### 2021-05-05 6c374ca
trying again with the equation figure
