### 2021-05-05 53383ea
trying again with the equation figure
