+++ paper.md
@@ -74,29 +74,29 @@ combinations =
 \begin{pmatrix}
 1\\
 2\\
-.\\
-.\\
-.
+\cdot\\
+\cdot\\
+\cdot
 \end{pmatrix} \times
 \begin{pmatrix}
 1\\
 2\\
-.\\
-.\\
-.
+\cdot\\
+\cdot\\
+\cdot
 \end{pmatrix} \times ...
 \begin{pmatrix}
 1\\
 2\\
-.\\
-.\\
-.
+\cdot\\
+\cdot\\
+\cdot
 \end{pmatrix} =
 \begin{cases}
-[1 1 ... 1]\\
-[1 1 ... 2]\\
-[2 2 ... 1]\\
-[2 2 ... 2]
+[1\text{for }1\text{ }\cdot\cdot\cdot1]\\
+[1\text{ }1\text{ }\cdot\cdot\cdot\text{ }2]\\
+[2\text{ }2\text{ }\cdot\cdot\cdot\text{ }1]\\
+[2\text{ }2\text{ }\cdot\cdot\cdot\text{ }2]
 \end{cases}
 \end{equation}
 
