### 2021-05-05 5d50dc4
trying again with the equation figure
