### 2021-05-05 8a342b8
trying again with the equation figure
