### 2021-05-05 af3ecd9
trying again with the equation figure
