+++ paper.md
@@ -48,7 +48,7 @@ Table 1: Constraints, i.e. the variables and their meanings, and some example va
 | *t<sub>s</sub>* | shift lengths                      | 8.33 |
 | *t<sub>W</sub>* | weekly working hours per worker    | 36.00 |
 
-Since each week also resembles a worker, the shift array can be set up as a matrix with 7 columns (each representing the days of a week) and *n<sub>W</sub>/7* rows (each representing a worker). The columns can then be summed to achieve the shift occupancy (or how many people are working each shift). Thus, the phase1 algorithm only allows shift arrays to pass for which all shifts are occupied by at least one worker, with a shift represented by the first *n<sub>wd</sub>* days for each week. In order to extend to not only use single shifts but also 2- or 3-shifts, a logical condition was added into the algorithm: For *N* shifts per day, each day has to be filled with at least *N* workers.
+Since each week also resembles a worker, the shift array can be set up as a matrix with 7 columns (each representing the days of a week) and $$n_{W}/7$$ rows (each representing a worker). The columns can then be summed to achieve the shift occupancy (or how many people are working each shift). Thus, the phase1 algorithm only allows shift arrays to pass for which all shifts are occupied by at least one worker, with a shift represented by the first *n<sub>wd</sub>* days for each week. In order to extend to not only use single shifts but also 2- or 3-shifts, a logical condition was added into the algorithm: For *N* shifts per day, each day has to be filled with at least *N* workers.
 
 In order to avoid all working days from being clustered together, the constraint for weekly minimum single continuous resting time is added (*t<sub>r</sub>*). The algorithm ensures that all passed shift arrays have at least *t<sub>r</sub>* hours of free-time over any given 7-day period.
 
