### 2021-05-05 b32e4e4
trying again with the equation figure
