### 2021-05-05 c918a41
trying again with the equation figure
