### 2021-05-05 e871582
trying again with the equation figure
