### 2021-05-05 525bcec
trying again with the equation figure
