### 2021-06-12 #3246 @majensen
@amoeba would you have time/interest in review [this submission](#3246) for JOSS? Appreciate the consideration as always.
