### 2021-06-21 #3246 @majensen
@sap218 would you be willing to review [this work](#3246) for JOSS? I would appreciate your expertise here.
Thanks for the consideration! --@majensen
