### 2021-06-21 #3246 @benjaminbolling
My initial suggestions after looking through the editorial board would be one of following:
- @melissawm (applied mathematician and an editor for Python scripts)
- @lorenanicole (data engineer and data scientist, a champion of FATE and an editor for Python scripts)
- @jbytecode (Author of three books on R programming and genetic algorithms, and developer and maintainer of many open source projects in several languages including R, Julia, Python, and Java)
