### 2021-06-21 #3246 @danielskatz
@benjaminbolling  - please don't suggest editors as reviewers; they already have enough work 😄

As noted in the first comment in this issue, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

And also, as mentioned in that comment, please suggest people without tagging them with an @
