### 2021-06-22 #3246 @majensen
@magedhelmy1 @ShantanuDash Thanks very much for the interest. Please have a look at the process here https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and here https://joss.readthedocs.io/en/latest/review_criteria.html - if you have any questions just ask here. If you change your mind, that is ok!
