### 2021-06-29 #3431 @whedon
[REVIEW]: Computer-Aided Generation of N-shift RWS
**Submitting author:** @benjaminbolling (<a href="http://orcid.org/0000-0002-6650-5365">Benjamin Edward Bolling</a>)
**Repository:** https://github.com/benjaminbolling/RSW
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
