### 2021-07-05 50b3ed4
Motivation updated
