### 2021-07-05 9bbafce
README file: added licencing and contributing sections. Preparing a test protocol file
