### 2021-07-05 3060ddc
README file: added licencing and contributing sections. Preparing a test protocol file
