+++ paper.md
@@ -26,7 +26,7 @@ bibliography: docs/paper.bib
 # Statement of Need
 All around the world, research institutes and industrial complexes make use of a workforce working multiple shifts per day in order to utilise maximum efficiency and profitability of the facility. Creating shift work schedules has, however, always been a challenging task, especially such that are equal for all workers and at the same time distributes the shifts evenly and properly to prevent staff burnout.
 
-The purpose and aim of this package is hence to support research institutes and industrial complexes at which non-standard working hours are applicable with a computational tool to create rotational workforce schedules by providing the user (schedule-maker) with all possible schedules for a set of input conditions, including parameters such as shift lengths, weekly working hours and -resting time, etc., by utilising a Combinatoric Generator and a Cartesian Product calculator.
+The purpose and aim of this package is hence to support research institutes and industrial complexes at which non-standard working hours are applicable with a computational tool to create rotational workforce schedules by providing the user (schedule-maker) with all possible schedules for a set of input conditions (such as shift lengths, weekly working hours and -resting time) by utilising a Combinatoric Generator and a Cartesian Product calculator.
 
 ## Introduction
 In order to achieve schedules for the workers that treats everyone equally, the focus of this project is on so-called rotational workforce schedules (RWSs). Rotational workforce schedules means that the schedule rotates after time, and hence, the other option would be static shift schedules. In this project, the term 'shift arrays' is defined to represent all possible schedules following a list of constraints, originating from e.g. country laws and/or workforce requests.
