### 2021-07-05 753a429
Updated requirements, environment and documentation for the added .ODS-export function
