+++ README.md
@@ -69,7 +69,7 @@ Table 1: Constraints, i.e. the variables and their meanings, and some example va
 
 After defining these values, the *Generate Combinations* button is clicked. During the generation, the approved combinations are stored in a temporary file. When finished, the generated shift arrays (combinations) can be browsed through using the slider or the numerical input field. The shift arrays are constructed such that each week or worker (depending on the viewing angle) is represented in separate rows.
 
-The slider is moved to combination \#212, and the bottom-right button *Proceed with this combo to next phase* is clicked - launching phase 2 on this combination as input. In the phase2 window, the button *Find Solutions* is clicked, generating all possible solutions for the combination. The slider in this window is not moved, making it remain on solution \#0. The solution is exported as a .CSV file. The same is repeated but using combination \#43 in phase1 and solution \#43 in phase 2. The files are imported into a new spreadsheet in Microsoft Excel (2018) and combined, followed by adding colours to highlight and differentiate each week, and lines for all rows and columns. The sum of each shift specie per day has also been calculated in the spreadsheet (using the COUNTIF(cells;shift) formula of Microsoft Excel (2018)), see bottom of Figure 3.
+The slider is moved to combination \#212, and the bottom-right button *Proceed with this combo to next phase* is clicked - launching phase 2 on this combination as input. In the phase2 window, the button *Find Solutions* is clicked, generating all possible solutions for the combination. The slider in this window is not moved, making it remain on solution \#0. The solution is exported as a .CSV file. The same is repeated but using combination \#43 in phase1 and solution \#43 in phase 2. The files are imported into a new spreadsheet in Microsoft Excel (2018) and combined, followed by adding colours to highlight and differentiate each week, and lines for all rows and columns. The sum of each shift specie per day has also been calculated in the spreadsheet (using the COUNTIF(cells;shift) formula of Microsoft Excel (2018)), see bottom of Figure 3. Another schedule was exported directly as a .ODS-file (OpenDocument Spreadsheet), in which the sum of each shift specie per day is included in the exported file with a screenshot attached in Figure 4.
 
 These two schedules are filled as evenly as possible with the exception of Fridays, which have been selected such that all shift workers are scheduled to work (suitable for e.g. activities when *all hands on deck* is required. This schedule has also ensured that each shift is occupied by a minimum of two shift workers.
 
@@ -85,6 +85,9 @@ Figure 2: The RWS:ing Application's algorithm's "phase 2 GUI" as launched from t
 
 ![Figure 3](docs/fig3.png)
 Figure 3: Example procedure results. Phase 1 was performed using the parameters defined in Table 1 and with free days clustered in pairs. Two boolean combinations were selected, followed by finding and selecting a solution for both, and then exported as CSV and imported in Microsoft Excel (2018).
+
+![Figure 4](docs/fig5.png)
+Figure 4: A combination for 2-shifts was put together but not solved (for demo-purposes) showing the shift arrays and the occupancy per shift and day.
 
 ## License
 The `RSW` package is licensed under the GNU General Public License v3.0.
