+++ requirements.txt
@@ -1,2 +1,3 @@
 numpy==1.19.2
+pyexcel==0.6.6
 PyQt5==5.15.4
