+++ environments.yml
@@ -5,4 +5,5 @@ channels:
 dependencies:
   - python
   - numpy
+  - pyexcel
   - pyqt
