### 2021-07-06 bf3ab74
Possibility to save 1 combo from phase 1. Bug fix for loading 1-shift arrays in phase 2.
