### 2021-07-06 f3bf15f
Put the previous fig1 back
