+++ README.md
@@ -78,7 +78,7 @@ These two schedules are filled as evenly as possible with the exception of Frida
 Figure 0: The RWS:ing Application's launcher.
 
 ![Figure 1](docs/fig1.png)
-Figure 1: The RWS:ing Application's algorithm's "phase 1 GUI".
+Figure 1: The RWSing Application's algorithm's "phase 1 GUI" (dark and light themes, left and right, respectively). In the left figure, the combinations have been generated. In the right figure, the combinations have been loaded from a file.
 
 ![Figure 2](docs/fig2.png)
 Figure 2: The RWS:ing Application's algorithm's "phase 2 GUI" as launched from the "phase 1 GUI" and with the second Thursday's shift changed to an evening shift (left) and after finding solutions, showing the first solution (right).
