### 2021-07-06 2c8c7ef
added python and pyqt5 references in paper.md
