+++ docs/paper.bib
@@ -1,13 +1,3 @@
-@article{EfficientRWS,
-  author =        {N. Musliu, J. Gärtner, and W. Slany},
-  journal =       {Discrete Applied Mathematics},
-  pages =         {85--98},
-  title =         {{Efficient generation of rotating workforce schedules}},
-  volume =        {118},
-  year =          {2002},
-  doi =           {10.1016/s0166-218x(01)00258-x}
-}
-
 @webpage{PythonReference,
     Author = {{Riverbank Computing Limited}},
     title = {{Python Language Reference}},
