### 2021-07-06 c0d2283
testing decreasing picture width
