### 2021-07-06 a6a8904
Fixed a bug : If 0 is first shift, phase2 gets in trouble
