+++ docs/paper.bib
@@ -1,5 +1,5 @@
 @webpage{PythonReference,
-    Author = {{Riverbank Computing Limited}},
+    Author = {{Python Software Foundation}},
     title = {{Python Language Reference}},
     URL = {http://www.python.org},
     version = {3.8.2},
