### 2021-07-16 benjaminbolling/RSW#1 @ShantanuDash
Update `Requirements.txt`
The `Requirements.txt` should have atleast a range of version support or support newer versions.

openjournals/joss-reviews#3431
