### 2021-07-16 benjaminbolling/RSW#2 @ShantanuDash
Errors due to path problems
Having spaces in the folder names seems to cause problems

```
D:\joss review\RSW>python RSW.py
python: can't open file 'D:\joss': [Errno 2] No such file or directory
```

Replacing space with hyphen removes the problem.

```
D:\joss-review\RSW>python RSW.py

D:\joss-review\RSW>
```

Possible solutions:
- https://stackoverflow.com/questions/50011733/python-error-filenotfounderror-errno-2-no-such-file-or-directory/50011910
- https://stackoverflow.com/questions/12201928/python-open-gives-filenotfounderror-ioerror-errno-2-no-such-file-or-directory

openjournals/joss-reviews#3431
