### 2021-07-16 a0eb31e
Requirements updated : Ranges for PyQt5 and numpy
