### 2021-07-16 benjaminbolling/RSW#1 @benjaminbolling
Done for PyQt5 and and NumPy. PyExcel will however have to remain in a specific versioned state.
