### 2021-07-16 benjaminbolling/RSW#2 @ShantanuDash
Joss review folder has the repo inside it.
So the problem arises due to the reasons given in the Stackoverflow I've linked above.

If the changes seem too cumbersome then you could add a suggestion for users _to not have space separated directories_ .
