### 2021-09-24 197a958
WIP on the paper. SW high level architecture flow has been added.
