+++ docs/paper.bib
@@ -5,10 +5,50 @@
     version = {3.8.2},
     year = {2020}
 }
-
 @webpage{PyQtReference,
     Author = {{Riverbank Computing Limited}},
     title = {{PyQt5}: {P}ython bindings for the {Q}t cross platform {UI} and application toolkit},
     URL = {https://www.riverbankcomputing.com/software/pyqt/},
     year = {2016}
+}
+@article{Laporte2004,
+    title = {A General Multi-Shift Scheduling System},
+    author = {Gilbert Laporte and Gilles Pesant},
+    journal = {The Journal of the Operational Research Society},
+    volume = {55},
+    pages = {1208–1217},
+    year = {2004}
+    doi = {10.1057/palgrave.jors.2601789},
+}
+@article{Becker2020,
+    title = {A decomposition heuristic for rotational workforce scheduling},
+    author = {Tristan Becker},
+    journal = {Journal of Scheduling},
+    volume = {23},
+    pages = {539–554},
+    year = {2020}
+    doi = {10.1007/s10951-020-00659-2},
+}
+@article{Burke2004,
+    title = {The state of the art of nurse rostering},
+    author = {Burke, E., De Causmaecker, P., Berghe, G. V., & Van Landeghem, H.},
+    journal = {Journal of Scheduling},
+    volume = {7(6)},
+    pages = {441–499},
+    year = {2004}
+    doi = {10.1023/B:JOSH.0000046076.75950.0b},
+}
+@inproceedings{Musliu2011,
+    author = {Musliu, N.},
+    title = {Constructing cyclic staff schedules by iterated local search},
+    booktitle = {In MIC 2011: The IX Metaheuristics International Conference},
+    year = {2011},
+    pages = {3-5},
+}
+@inproceedings{PATAT2002,
+  author = {Petrovic, S., G. Beddoe, and G. Vanden Berghe},
+  title = {Storing and adapting repair experiences in personnel rostering},
+  booktitle = {Practice and Theory of Automated Timetabling, Fourth International Conference},
+  year = {2002},
+  pages = {185-186},
 }
