+++ docs/paper.bib
@@ -11,6 +11,21 @@
     URL = {https://www.riverbankcomputing.com/software/pyqt/},
     year = {2016}
 }
+@misc{EncyMath,
+    author = {{Springer Verlag GmbH, European Mathematical Society}},
+    title = {{Encyclopedia of Mathematics}},
+    howpublished = {Website},
+    note = {URL: \url{https://www.encyclopediaofmath.org/}. Accessed on 2021-11-20}
+}
+@article{Tristan2020,
+    Author = {Tristan Becker},
+    title = {A decomposition heuristic for rotational workforce scheduling},
+    journal = {Journal of Scheduling},
+    volume = {23},
+    pages = {539-554},
+    year = {2020}
+    doi = {10.1007/s10951-020-00659-2},
+}
 @article{Laporte2004,
     title = {A General Multi-Shift Scheduling System},
     author = {Gilbert Laporte and Gilles Pesant},
@@ -19,15 +34,6 @@
     pages = {1208–1217},
     year = {2004}
     doi = {10.1057/palgrave.jors.2601789},
-}
-@article{Becker2020,
-    title = {A decomposition heuristic for rotational workforce scheduling},
-    author = {Tristan Becker},
-    journal = {Journal of Scheduling},
-    volume = {23},
-    pages = {539–554},
-    year = {2020}
-    doi = {10.1007/s10951-020-00659-2},
 }
 @article{Burke2004,
     title = {The state of the art of nurse rostering},
