+++ docs/paper.bib
@@ -23,7 +23,7 @@
     journal = {Journal of Scheduling},
     volume = {23},
     pages = {539-554},
-    year = {2020}
+    year = {2020},
     doi = {10.1007/s10951-020-00659-2},
 }
 @article{Laporte2004,
