+++ docs/paper.bib
@@ -3,19 +3,19 @@
     title = {{Python Language Reference}},
     URL = {http://www.python.org},
     version = {3.8.2},
-    year = {2020}
+    year = {2020},
 }
 @webpage{PyQtReference,
     Author = {{Riverbank Computing Limited}},
     title = {{PyQt5}: {P}ython bindings for the {Q}t cross platform {UI} and application toolkit},
     URL = {https://www.riverbankcomputing.com/software/pyqt/},
-    year = {2016}
+    year = {2016},
 }
 @misc{EncyMath,
     author = {{Springer Verlag GmbH, European Mathematical Society}},
     title = {{Encyclopedia of Mathematics}},
     howpublished = {Website},
-    note = {URL: \url{https://www.encyclopediaofmath.org/}. Accessed on 2021-11-20}
+    note = {URL: \url{https://www.encyclopediaofmath.org/}. Accessed on 2021-11-20},
 }
 @article{Tristan2020,
     Author = {Tristan Becker},
@@ -32,7 +32,7 @@
     journal = {The Journal of the Operational Research Society},
     volume = {55},
     pages = {1208–1217},
-    year = {2004}
+    year = {2004},
     doi = {10.1057/palgrave.jors.2601789},
 }
 @article{Burke2004,
@@ -41,7 +41,7 @@
     journal = {Journal of Scheduling},
     volume = {7(6)},
     pages = {441–499},
-    year = {2004}
+    year = {2004},
     doi = {10.1023/B:JOSH.0000046076.75950.0b},
 }
 @inproceedings{Musliu2011,
