+++ paper.md
@@ -46,7 +46,7 @@ Table 1: Constraints, i.e. the variables and their meanings, and some example va
 | Variable       | Meaning     | Value |
 | :------------- | :----------: | :----------: |
 | $N$ | number of shifts per days            | 2 |
-| $n_{cf}$ | number of days off clustered      | 2 |
+| $n_{cf}$ | number of days off clustered      | - |
 | $n_{S}$ | number of shifts per shift cycle      | 18 |
 | $n_{W}$ | number of weeks to cycle over      | 4 |
 | $n_{wd}$ | number of working days per week   | 7 |
@@ -126,24 +126,9 @@ Pressing the *Find solutions* results in what is shown in \autoref{fig:phase2} (
 
 where each array in the resulting product is considered as a possible shift schedule matrix. Imposing constraints (resting time between shifts and ensuring all shifts are filled) on each combinations results in solutions from which the user can choose between.
 
-Since all combinations are stored in a matrix form before different combinations are removed from the final solutions matrix, large datasets require severe amount of internal memory for the Cartesian Product method to work. For this, a controlling script has been implemented which calculates a pre-estimate of required internal memory. Since the array consists of integers (zeroes and ones) and an integer in Python occupies 4 bytes, a list occupies 64 bytes with 8 byte per item within, a simple expression can be created as follows:
-
-\begin{equation}
-IM \approx 64 + 12 \times (N_{int})
-\end{equation}
-with $N_{int}$ being the total number of integers from all combinations combined.
-
-This means that the required internal memory for different operations can be roughly calculated by
-\begin{equation}
-IM \approx 64 + 12 \times N_{C} \times n_{S} = 64 + 12 \times N^{n_{S}} \times n_S
-\end{equation}
-returning the memory demand IM in bytes and where $N_{C} = N^{n_{S}}$ is the total number of combinations (without any constraints imposed).
-
-If the estimated expected internal memory requirement for an operation exceeds 1Gb, the user is prompted whether to continue with the default Cartesian Product method or to use a less internal memory demanding recursive method.
-
 ## Benchmarking results
 ### Benchmarking Computer Specifications
-The algorithm benchmarking was done on an Apple MacBook Pro with the specifications as defined in Table 2.
+The algorithm benchmarking was done on an Apple MacBook Pro with the specifications defined in Table 2.
 
 Table 2: Benchmarking computer specifications.
 
@@ -160,14 +145,14 @@ In the GUI, there is a "fast generation" checkbox which stops the algorithm from
 
 The parameters used are defined in Table 1, with the exception of $N$ and Shift types' labels. Note that for Table 3, the number (\#) of weeks given is the minimum amount of weeks required for a full shift cycle in order to find acceptable combinations for the N-shift problems (with $N = 1,2,3$ for single-, two- and three-shifts, respectively). The free days clustering option is not selected for the benchmarking.
 
-Table 3: Benchmarking for fast and full generation of the Boolean Arrays (as defined in Section 3.1 for Phase 1), and the number of combinations and approved combinations found for full generations of the Boolean Arrays (as defined in Section 3.1 for Phase 1). The types are single-, two- or three-shifts during 5 or 7 days per week.
+Table 3: Benchmarking for fast and full generation of the Boolean Arrays (as defined in Section 3.1 for Phase 1), and the number of combinations and approved combinations found for full generations of the Boolean Arrays (as defined in Table 1). The types are single-, two- or three-shifts during 5 or 7 days per week.
 
[…]
