### 2021-11-21 8085b78
Table 4 values updated
