### 2022-01-12 #3431 @magedhelmy1
Hi Benjamin, thanks for updating the paper.
It looks much better now.

I have now ticked the parts related to the paper.

- Reference
- Summary
- A statement of need
- State of the field
---------------------------
I have follow-up feedback/questions.
1. Functionality: Have the functional claims of the software been confirmed?

1.1 - I am finding it hard to follow the software's functionality from just screenshots and captions. Can you please include a [user flow](https://userguiding.com/blog/ux-user-flow-diagram/) of all the scenarios possible with your software, so it is clear how to use it? There is no one way of doing a user flow so, please do whatever you are most confident with. The user flow gives a nice bird's-eye view of your software.
1.2 - For a non-expert on work shifts, I find it hard to understand the labels of the fields. For example, "Cluster Free Days" - "Number of Weeks to Cycle Over," etc. Can you please include a document explaining these fields?
1.3 - The input fields can only be controlled by the arrows, for example, _shift lengths [h]_ can only go up xx.33 but not for example xx.66. Is this on purpose to only use .33?
1.4 - I see "Load Combos" but "generate combinations", is combos, combinations? If so, kindly use a standardized language across.
1.5 - When I input a great workload, I get the message "10 weeks might require an extensive amount of computing power...", and the software becomes "non-responding"; as you can see my computer CPU and RAM are quite relaxed. I am not sure why that happens. Can you please elaborate on why is this happening although there are available computing resources?

I will probably come with more questions after seeing the user flow diagrams since I could not navigate and understand the full software with just images and the caption.

2. Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
2.1 - I am not sure I find this anywhere. Can you please point me to it?
2.2 - To add, can you please at least provide Level 1 and 2 diagrams from the [C4 model ](https://c4model.com/) to understand better the purpose of the functions and classes you have created?

3. Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

This question goes back to @majensen as he is the moderator.

Specifically to these 2 points from the JOSS author guidelines:

1. "Your software should be a significant contribution to the available open-source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).."

2. "Whether the software is sufficiently useful that it is likely to be cited by your peer group."

I am not sure there is a specific research challenge that has been addressed per se.
As stated by the author, this is a  "Multi-Shift Rotational Workforce Schedules" generator. This is more of a nice to have "App" to me. I can find tons of them [online](https://peoplemanagingpeople.com/tools/best-employee-shift-scheduling-software/). So although this can be a good "app", it is not necessarily a research app. This brings me to the second point stated by the guidelines, and I think the answer to the next question is "No". But @benjaminbolling I am happy to hear your comments about this part and if you can highlight the research parts of the app? Thank you.
