### 2022-01-13 #3431 @benjaminbolling
Thanks for your comment @magedhelmy1, a lot of things to address I see. :)

For point 3:
So the difference between this software and many apps available online is that this is a semi-automised method for generating shift patterns, whilst the ones shown in the list above allows more-or-less users (schedule-designers/makes) to insert and plan manually.

The reason why I would consider this as a research software is that it uses a computational approach to construct initial rotational shift work schedule patterns, and allows for implementation of further algorithms that can be developed and inserted (to be invoked e.g. under ` def createAllShiftPossibilities` ). Hence my assumption that it can address the research challenge of finding appropriate shift patterns equal for all employees.

Please let me know if this clarifies @magedhelmy1 , or if this maybe is not considered as a research app - to be clarified by @majensen .

I will address points 1+2 with their respective sub-points later today.
