### 2022-01-13 #3431 @magedhelmy1
Hi Benjamin, Thanks for the reply.

The sentence you wrote under point 3 above should be reflected in the State-of-the-art part in your paper.
As the guidelines of JOSS state.

"..enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).."

So kindly update the paper to clearly reflect this point and then you can cite those other software collections. Just a suggestion, maybe a simple table of the other software name, what they do, and how your software does it better :) But you are the author so it is up to you.
