### 2022-01-19 #3431 @majensen
@magedhelmy1 thanks for picking this up again. You raise a fair point, and I would like to see @benjaminbolling address them with some specifics in the paper. I will then take a look, and may ask some other editors for their input.

One approach, @benjaminbolling, may actually be to talk about your specific application in your work-- maintenance of scientific megainstruments round the clock so experiments don't fail. Can you highlight any specific problems that your software solves in this regard that are different from run-of-the-mill scheduling?
