+++ README.md
@@ -13,9 +13,7 @@ In this project, an algorithm takes into account inputs (or constraints) from th
 ## Getting started
 A suggestion on how to get started is given in a chronologically suggested order:
 - Installation procedure
-- Description of how to launch the package
-- Description of how to generate combinations fulfilling user requirements
-- Description of how to solve different combinations
+- Description of how to use the package
 - Example usage procedure
 
 ### Installation procedure
@@ -31,31 +29,42 @@ A suggestion on how to get started is given in a chronologically suggested order
 
 4. If all required Python packages have been successfully installed, the package is ready.
 
+## Description of software
+### Nomenclature
+In this application, the following terminology is used:
+- The problem: The rotational shift-work schedule that needs to be created based upon a set of constraints.
+- Combination: A list of zeroes (days off) and ones (work days)
+- Solution: A combination that solves the problem.
+
 ### Launcher
 The package includes a simple launcher which can be launched by executing the following (with current directory being this repository):
 
     python RSW.py
 
-The launcher gives the user to either start from scratch (phase 1), continue from a saved file in phase 1 (a list of combinations), or continue from a saved file in phase 2 (the solutions). See respective sections Phase 1 and Phase 2 below for descriptions of each.
+The launcher gives the user to either start from scratch, continue from a saved file in "Phase 1" (from a previously generated list of combinations), or continue from a saved file in "Phase 2" (potential solutions).
 
 ### Phase 1
-Phase 1 consists of a list of parameters that are to be filled in such that schedules can be generated (by clicking on the *Generate Combinations* button). Note that the larger the number of weeks is, the longer it will take to generate the combinations.
+Phase 1 consists of creating boolean combinations that may solve the problem, see Figure 1 for the GUI.
 
-When the combinations have been generated and at least one is approved (fulfilling the constraints), the combinations can be browsed using the combination index spinbox or -slider. Each column represents a workday and each row a worker for week \#0, with booleans showing if a worker is working (1) or is free (0).
+Upon filling in the list of constraints, pressing the "Generate Combinations" button activates the algorithm that runs through the various permutations of boolean lists. The algorithm returns a list of combinations, of which each have the potential of solving the problem.
 
-The combinations can be tested if they have any real solution in phase 2 by clicking the bottom-left button *Test For Solutions*. Note: This procedure can take long time to complete, with time for completion increasing with number of shifts and number of weeks in a cycle.
+Each combination is presented one-by-one in a simple matrix where each column is a day (e.g. Mon-Fri or Sun-Sat). The slider and spin-box next to it allows for quick browsing through the different combinations.
+
+The button "Proceed with this combo to next phase" launches "Phase 2" with the combination.
 
 ### Phase 2
-The "phase 2 GUI" is launched with a selected combination as input (which can be seen in Figure 2 to the left).
+Phase 2 consists of finding solutions to a combination that can solve the problem, see Figure 2 for the GUI.
 
-The free days are all represented by zeroes whilst all other shifts (ones) are converted to the first defined shift type label. For *N* > 1, each shift can be replaced by another shift via dropdown menus. The GUI shows the number of shifts of each type per week and a table with the results, i.e. the number of workers per shift and day. Shifts that are occupied have green background whilst shifts that are unoccupied have a red background.
+It can be launched with a saved file from the launcher or from Phase 1.
 
[…]
