### 2023-01-28 #3431 @benjaminbolling
Points 1.1-1.4 have been resolved in commit https://github.com/benjaminbolling/RSW/commit/cd615cbcbb274e3f5cf59c6ccac8bd704f6ff86c.

Point 1.5 will be solved by making some optimisation in the code, that will unfortunately need a little bit more time (which I am happy to state I will have in the coming weeks).
