### 2023-01-28 #3431 @benjaminbolling
Point 3 hopefully considered as resolved in commit https://github.com/benjaminbolling/RSW/commit/cd615cbcbb274e3f5cf59c6ccac8bd704f6ff86c.
