### 2023-01-28 #3431 @benjaminbolling
Point 2 hopefully resolved in https://github.com/benjaminbolling/RSW/commit/1ee4c61e9a8a3f8d4ea3c09adde8eb3174da6863.
