### 2023-01-28 7527669
Changed a figure size
