### 2023-01-28 9328e3a
Changed a figure size
