### 2023-01-28 b17bbbd
Changed a figure size
