### 2023-01-29 #3431 @majensen
Hi @benjaminbolling - thanks for putting in this work. I think it is looking good. Can you please consider the following small edits:

l. 31: "Conclusively, " - delete
l. 69: "with" -> "to have"
l. 69: "n_{cf}" didn't get formatted (maybe an unusual underline character?)
l. 83: remove the colons (i.e., say 1s, 0s, or 'ones', 'zeroes')
l. 84: No need to cite a reference the common expression for 'n choose r' - you can remove this
l. 151: Citation should look like "In 2004, Burke (2004) made ...". This can be done with Latex, but can't remember how off hand.
l. 157 Citation has double parentheses - should be single
l. 162 (Petrovic & Berghe, 2002) -> Petrovic & Berghe (2002)
l. 162 "includes and" - delete
l. 164 (Laporte & Pesant, 2004)  -> Laporte & Pesant (2004)
l. 168-169 "does not give the user ... middle" -> "does not allow the user to perform the middle step"
