### 2023-01-29 #3431 @majensen
@benjaminbolling - The associate editor-in-chief may think the paper as a whole is too long. Would you be willing to put the Benchmarking section in a readable form (markdown, eg) on your GitHub repo, and refer out to it in your paper?
