### 2023-01-30 f131259
Comments from majensen taken into account (https://github.com/openjournals/joss-reviews/issues/3431#issuecomment-1407770381), left to do is to fix the citation format.
