+++ paper.md
@@ -32,7 +32,7 @@ Comparing to already available software that (by the developer/author) could be 
 
 Another approach and need of the application is to serve as a support function for creating rotational shift workforce schedules for large-scale scientific research facilities such as particle accelerators, at which the workforce size is many time very constrained. In those situations, the creation of schedules can be very complex and difficult, making a tool such as this useful. In other cases, it can be used to prove that the constraints are too tight and that these cannot be fulfilled by any schedule, saving time for shift schedulers and hence also for the research facility.
 
-Conclusively, this package provides a graphical user interface (based on PyQt5 [@PyQtReference]) tool for generating and constructing acceptable shift arrays if there are any possible arrays following all user-defined constraints. These can be exported to the file formats ODS, CSV, and txt, with the arrays ready to be used as they are or as templates for further modifications (e.g. swapping shifts between workers and hence taking into account individual workers' needs).
+This package provides a graphical user interface (based on PyQt5 [@PyQtReference]) tool for generating and constructing acceptable shift arrays if there are any possible arrays following all user-defined constraints. These can be exported to the file formats ODS, CSV, and txt, with the arrays ready to be used as they are or as templates for further modifications (e.g. swapping shifts between workers and hence taking into account individual workers' needs).
 
 \pagebreak
 
@@ -70,13 +70,13 @@ The number of shifts per shift array is, in this algorithm, calculated by
 n_{S} = \text{ceil}(t_W / t_s)
 \end{equation}
 
-with the reason for using ceiling function (and not the floor function) being the argument that it is better with a couple of more hours than fewer. In order to cluster days off (n_{cf}), the algorithm's GUI has an optional additional constraint that serves this purpose and simply does not allow shift arrays with 0:s in clusters less than this through.
+to have the reason for using ceiling function (and not the floor function) being the argument that it is better with a couple of more hours than fewer. In order to cluster days off ($n_{cf}$), the algorithm's GUI has an optional additional constraint that serves this purpose and simply does not allow shift arrays with zeroes in clusters less than this through.
 
 By using the input $n_W \times n_{wd}$ as the iterable and $n_{S}$ as the length of subsequences of elements from the iterable, the same methodology as the *combinations* function of the *itertools* module in Python [@PythonReference] (a combinatoric generator) is used for creating each shift array. It can be simply described as creating an array of combinations (in this case, zeroes and ones corresponding to a day off or shift work, respectively) with a specific length (number of days in a cycle). By imposing the other inputs as constraints on whether a shift array should be appended to accepted shift arrays, the reason for not using the built-in Python module becomes clear: Python's built-in module returns all array combinations that are possible without any imposed constraints, which quickly escalates to becoming too large for a personal computer's internal memory to handle.
 
-With this, the final result is an array of shift arrays in which each shift array is filled with $7n_{S}$ 1:s and $n_{W}(7-n_{S})$ 0:s whilst obeying the above mentioned constraints.
+With this, the final result is an array of shift arrays in which each shift array is filled with $7n_{S}$ ones and $n_{W}(7-n_{S})$ zeroes whilst obeying the above mentioned constraints.
 
-As there are ${n\choose r}$ ways to choose r elements from a set of n elements [@EncyMath], the number of possible combinations ($C$) can be expressed by using the factorial of the binomial coefficient:
+As there are ${n\choose r}$ ways to choose r elements from a set of n elements, the number of possible combinations ($C$) can be expressed by using the factorial of the binomial coefficient:
 
 \begin{equation}
 C = \frac{n!}{r! \times (n-r)!}.
@@ -153,7 +153,7 @@ Table 2: Benchmarking computer specifications.
 | Graphics Card: |  Intel Iris Plus Graphics 655 1536 MB |
 
 ### Benchmarking Phase 1
[…]
