+++ paper.md
@@ -138,62 +138,6 @@ IM \approx N_{size} = N^{n_{S}} \times n_S,
 
 where $N_{size}$ is the total number of zeroes and ones in the full matrix. If the estimated size of the resulting matrix from the operation exceeds 1Gb, the user is prompted whether to continue with the default Cartesian Product method or to use a less internal memory demanding recursive method.
 
-## Benchmarking results
-### Benchmarking Computer Specifications
-The algorithm benchmarking was done on an Apple MacBook Pro with the specifications defined in Table 2.
-
-Table 2: Benchmarking computer specifications.
-
-| Definition     | Value        |
-| :------------- | :----------: |
-| Computer type: | Apple MacBook Pro (13-inch, 2019) |
-| OS: | macOS Mojave v. 10.14.6 |
-| Processor: |  2.8 GHz Intel Core i7 processor |
-| Internal Memory: | 16 GB 2133 MHz LPDDR3 |
-| Graphics Card: |  Intel Iris Plus Graphics 655 1536 MB |
-
-### Benchmarking Phase 1
-In the GUI, there is a "fast generation" checkbox which stops the algorithm from further calculations once the first 100 approved combinations have been found. This way, computation time can be lowered (in comparison to "full generation" which will go through all possible combinations from the boolean array). For the parameters defined in Table 1, the time it took to complete decreased from 508.7 s to 24.55 s for the full generation (see Table 3), which is a decrease in time by 95\%.
-
-The parameters used are defined in Table 1, with the exception of $N$ and Shift types' labels. Note that for Table 3, the number (\#) of weeks given is the minimum amount of weeks required for a full shift cycle in order to find acceptable combinations for the N-shift problems (with $N = 1,2,3$ for single-, two- and three-shifts, respectively). The free days clustering option is not selected for the benchmarking.
-
-\pagebreak
-
-Table 3: Benchmarking for fast and full generation of the Boolean Arrays (as defined in Section 3.1 for Phase 1), and the number of combinations and approved combinations found for full generations of the Boolean Arrays (as defined in Table 1). The types are single-, two- or three-shifts during 5 or 7 days per week, and the internal memory (IM) is given for the different shift array objects.
-
-| Type: | 1-shift, 5d/w | 1-shift, 7d/w | 2-shift, 7d/w | 3-shift, 7d/w   |
-| :------------- | :----------: | :----------: | :----------: | :----------: |
-| \# of weeks: | 1 | 2 | 4 | 5 |
-| Combinations (total): | 1 | 2 002 | 1.312e+07 | 1.476e+09 |
-| Combinations (accepted): | 1 | 462 | 1.668e+06 | 1.138e+07 |
-| IM: | 88 B | 4.216 kB | 13.53 MB | 100.4 MB |
-| Time (fast) [s]: | 7.224e-05 | 1.497e-02 | 24.55 | 3 087 |
-| Time (full) [s]: | 7.224e-05 | 5.211e-02 | 508.7 | 6.627e+04 |
-
-Plotting the benchmarking results yields the logarithmic graph in \autoref{fig:benchmarking}. As can be seen, the computation time $T_{C}$ increases exponentially with the number of weeks in a shift cycle on average in accordance with
-\begin{equation}
-T_C \text{(full)} = \exp{\{5.046 \times n_W\}} \times 9 \times 10^{-7}
-\end{equation}
-and
-\begin{equation}
-T_C \text{(fast)} = \exp{\{4.254 \times n_W\}} \times 2  \times 10^{-6}
-\end{equation}
-for the full and fast generations, respectively, calculated with an exponential regression.
-
-### Benchmarking Phase 2
[…]
