+++ docs/paper.bib
@@ -51,7 +51,7 @@
     year = {2011},
     pages = {3-5},
 }
-@inproceedings{PATAT2002,
+@inproceedings{PATAT:2002,
   author = {Petrovic, S., G. Beddoe, and G. Vanden Berghe},
   title = {Storing and adapting repair experiences in personnel rostering},
   booktitle = {Practice and Theory of Automated Timetabling, Fourth International Conference},
