+++ docs/paper.bib
@@ -17,7 +17,7 @@
     howpublished = {Website},
     note = {URL: \url{https://www.encyclopediaofmath.org/}. Accessed on 2021-11-20},
 }
-@article{Tristan2020,
+@article{Tristan:2020,
     Author = {Tristan Becker},
     title = {A decomposition heuristic for rotational workforce scheduling},
     journal = {Journal of Scheduling},
@@ -26,7 +26,7 @@
     year = {2020},
     doi = {10.1007/s10951-020-00659-2},
 }
-@article{Laporte2004,
+@article{Laporte:2004,
     title = {A General Multi-Shift Scheduling System},
     author = {Gilbert Laporte and Gilles Pesant},
     journal = {The Journal of the Operational Research Society},
@@ -35,7 +35,7 @@
     year = {2004},
     doi = {10.1057/palgrave.jors.2601789},
 }
-@article{Burke2004,
+@article{Burke:2004,
     title = {The state of the art of nurse rostering},
     author = {Burke, E., De Causmaecker, P., Berghe, G. V., & Van Landeghem, H.},
     journal = {Journal of Scheduling},
@@ -44,7 +44,7 @@
     year = {2004},
     doi = {10.1023/B:JOSH.0000046076.75950.0b},
 }
-@inproceedings{Musliu2011,
+@inproceedings{Musliu:2011,
     author = {Musliu, N.},
     title = {Constructing cyclic staff schedules by iterated local search},
     booktitle = {In MIC 2011: The IX Metaheuristics International Conference},
