### 2023-02-06 #3431 @majensen
:wave: @openjournals/dev  - can you please have a look at the error thrown while generating pdf at https://github.com/openjournals/joss-reviews/issues/3431#issuecomment-1418302724 ? Thanks
