### 2023-02-06 #3431 @xuanxu
The error is thrown because the author has no affiliation
