+++ paper.md
@@ -15,7 +15,11 @@ tags:
 authors:
   - name: Benjamin Edward Bolling
     orcid: 0000-0002-6650-5365
+    affiliation: "1"
 date: 25 September 2021
+affiliations:
+  - name: European Spallation Source ERIC
+    index: 1
 bibliography: docs/paper.bib
 ---
 
