### 2023-02-06 73b9c34
Moved acknowledgements back to after figures in accordance with https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography.
