### 2023-02-07 #3431 @benjaminbolling
Done: https://doi.org/10.5281/zenodo.7618915
