### 2023-02-08 #3431 @majensen
@benjaminbolling could you look at https://github.com/openjournals/joss-reviews/issues/3431#issuecomment-1423136022 and see if the reference recommedation makes sense , and if so add this to the bib?
