### 2023-02-08 #3431 @majensen
Note @benjaminbolling is updating references per editorialbot https://github.com/openjournals/joss-reviews/issues/3431#issuecomment-1423136022
> 👋 @openjournals/sbcs-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03431/joss.03431/10.21105.joss.03431.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#3953](https://github.com/openjournals/joss-papers/pull/3953), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`
