+++ docs/paper.bib
@@ -4,6 +4,7 @@
     URL = {http://www.python.org},
     version = {3.8.2},
     year = {2020},
+    doi = {10.1201/9781584889304-33},
 }
 @webpage{PyQtReference,
     Author = {{Riverbank Computing Limited}},
