### 2023-02-08 #3431 @benjaminbolling
Updated paper.bib for https://github.com/openjournals/joss-reviews/issues/3431#issuecomment-1423136022, see commit https://github.com/benjaminbolling/RSW/commit/91b09d016071baf422f3f793e164359269570bd0.
