### 2023-02-10 #3431 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
