### 2023-02-19 #3431 @arfon
@benjaminbolling – I found a minor issue with your paper. Could you please merge this PR? https://github.com/benjaminbolling/RSW/pull/10
